Use api-viewer and web-component-analyzer for easy docs viewing
