+++ demo/demo.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+    <script type="module" src="demo-element.js"></script>
+  </head>
+
+  <body>
+    <demo-element></demo-element>
+  </body>
+</html>
