+++ demo/index.html
@@ -1,11 +1,43 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
-    <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script type="module" src="demo-element.js"></script>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <title>Helium Animated Pages Docs</title>
+    <script type="module">
+      import 'api-viewer-element';
+    </script>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
+          Oxygen-Sans, Ubuntu, Cantarell, sans-serif;
+        margin: 0;
+      }
+      api-viewer::part(radio-button),
+      api-viewer::part(radio-label) {
+        display: none;
+      }
+      header {
+        position: absolute;
+        width: 100%;
+        max-width: 800px;
+        display: flex;
+        justify-content: flex-end;
+        padding: 0.9nrem;
+        box-sizing: border-box;
+      }
+      .demo {
+        color: #fff;
+        font-size: 0.875rem;
+        margin: 0px 0.75rem 0px 0.25rem;
+      }
+    </style>
   </head>
-
   <body>
-    <demo-element></demo-element>
+    <header>
+      <a class="demo" href="./demo.html">Demo</a>
+    </header>
+    <api-viewer src="../custom-elements.json"></api-viewer>
   </body>
 </html>
